+++ modules/client/test/integration/addresslistVoting-client/encoding.test.ts
@@ -10,9 +10,11 @@ import { bytesToHex, InvalidAddressError } from "@aragon/sdk-common";
 import { ADDRESS_ONE, contextParamsLocalChain } from "../constants";
 import { Context, SupportedNetworksArray } from "@aragon/sdk-client-common";
 
+const NETWORK_NAME = "goerli";
+
 jest.spyOn(SupportedNetworksArray, "includes").mockReturnValue(true);
 jest.spyOn(Context.prototype, "network", "get").mockReturnValue(
-  { chainId: 5, name: "goerli" },
+  { chainId: 5, name: NETWORK_NAME },
 );
 describe("Client Address List", () => {
   beforeAll(() => {
@@ -37,6 +39,7 @@ describe("Client Address List", () => {
       const installPluginItemItem = AddresslistVotingClient.encoding
         .getPluginInstallItem(
           withdrawParams,
+          NETWORK_NAME,
         );
 
       expect(typeof installPluginItemItem).toBe("object");